Stubbed nil values for the describe instance calls.
The default stub returns non-empty strings for all possible string
values. This causes the next token that trigger paging to appear
like there is a next response.

I updated your stubbs to use nil next tokens to avoid this issue.
Please feel free to open this as an issue against aws/aws-sdk-ruby as
this is a poor default experience.